Add files via upload
This notebook is very similar to the notebook "data_generator_ML_model", but here we are trying to find out if we could get similar results using only the minimum value of y for the beam, instead of all the values of y.  Some brief conclusions are in a comment in the end of the file.